+++ internal/model/Proposta.go
@@ -8,9 +8,9 @@ import (
 
 type Proposta struct {
 	BaseModel
-	IDVaga        	int64      `json:"id_vaga"`
+	IDVaga        	uint      `json:"id_vaga"`
 	Vaga			Vaga		`gorm:"foreignKey:IDVaga;references:ID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;" json:"vaga,omitempty"`
-	IDPrestador   	int64      `json:"id_prestador"`
+	IDPrestador   	uint      `json:"id_prestador"`
 	Prestador		Usuario		`gorm:"foreignKey:IDPrestador;references:ID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;" json:"prestador,omitempty"`
 	ValorProposto 	float64    `json:"valor_proposto"`
 	Mensagem      	string     `json:"mensagem"`
@@ -21,7 +21,7 @@ type Proposta struct {
 
 type PropostaRepo interface {
 	Criar(ctx context.Context, proposta *Proposta) error
-	ListarPorVaga(ctx context.Context, idVaga int64, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Proposta, error)
-	ListarPorPrestador(ctx context.Context, idPrestador int64, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Proposta, error)
-	AtualizarStatus(ctx context.Context, idProposta int64, status Status, dataResposta time.Time) error
+	ListarPorVaga(ctx context.Context, idVaga uint, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Proposta, error)
+	ListarPorPrestador(ctx context.Context, idPrestador uint, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Proposta, error)
+	AtualizarStatus(ctx context.Context, idProposta uint, status Status, dataResposta time.Time) error
 }
