+++ internal/model/transacao.go
@@ -4,9 +4,9 @@ import "context"
 
 type Transacao struct {
 	BaseModel
-	IDUsuario     int64        `json:"usuario_id" gorm:"not null"`
+	IDUsuario     uint        `json:"usuario_id" gorm:"not null"`
 	Usuario       *Usuario     `gorm:"foreignKey:IDUsuario;references:ID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;" json:"usuario,omitempty"`
-	IDServico     int64        `json:"servico_id" gorm:"not null"`
+	IDServico     uint        `json:"servico_id" gorm:"not null"`
 	Servico       *Servico     `gorm:"foreignKey:IDServico;references:ID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;" json:"servico,omitempty"`
 	TipoMovimento TipoMovimento `json:"tipo_movimento" gorm:"column:tipo_movimento;not null"`
 	Valor         float64      `json:"valor" gorm:"column:valor;not null"`
@@ -16,5 +16,5 @@ type Transacao struct {
 
 type TransacaoRepo interface {
 	Criar(ctx context.Context, transacao *Transacao) error
-		ListarPorUsuario(ctx context.Context, idUsuario int64, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Transacao, error)
+		ListarPorUsuario(ctx context.Context, idUsuario uint, filters map[string]interface{}, orderBy string, orderDir string, limit, offset int) ([]Transacao, error)
 }
